Reservation raises validationerrors when inproper data or field already reserved.
Issue @20
